(PUP-8010) Enable hash style iteration over non hash objects
This commit deals with the problem of deciding when to pass an index
or a key as the first argument to a two-element block that is passed to
an iterative function. Before this commit, the differntiation was made
solely based on if the iterable object was a `Hash` or something else
but with the introduction of other hash-like objects, that is no longer
enough.

The chosen solution was to add a `#hash_style?` method to the `Iterator`
class and a specialization to that class named `HashIterator`. All
`Iterator` implementations except the `HashIterator` will return false
on a call to `#hash_style?`.

The decision whether to use index or key as the first argument to the
block is now made based on the return value from `#hash_style?` and
iterators that are based on a `Hash` will use the `HashIterator`.