(PUP-8010) Add ability to implement an Iterable by producing an Iterator
This commit adds a new Ruby module named `IterableProducer`. It contains
one single method `#iterator` which is expected to produce an iterator
over the elements of an instance of the implementing class.

The `Iterable#on` method is made aware of this new module and when it
is encountered, the produced iterator will be returned.