(PUP-8012) Make it possible to add a '[]' function to an Object
This commit changes the constraints for the name of a function in an
`Object` declaration so that the special name '[]' is allowed. This
name is then recognized when the access operator is evaluated such that
an instance of `PuppetObject` with an `Object` type where such a
function is declared, will get that function invoked.

A '[]' function with a `RubyMethod` annotation will result in a Ruby
method with the `[]` name and hence, enable normal access op semantics
in Ruby.