(PUP-8128) Make it possible to pass pre-constructed types to dispatch
This commit makes it possible to pass pre-created type instances to
a Ruby function dispatch. Prior to this change, it was necessary to
pass a string that then would be parsed into a type. Creating such a
string for an already existing type was both cumbersome, error prone,
and easy to forget in case of subsequent changes to the type.