(PUP-8041) replace FACTER_url with target for puppet resource
Prior to this commit ...

It was necessary to set FACTER_url when using `puppet resource` with a device.
Doing so exposes device usernames and passwords on the command line:

`FACTER_url=https://admin@password@bigip.example.com puppet resource f5_vlan`

This commit ...

Defines a 'device()' method in network_device/config.rb.
Updates application/device.rb to use the 'device()' method.
Defines a '--target' option in application/resource.rb, resulting in:

`puppet resource --target bigip.example.com f5_vlan`

Notes

This change is backwards compatible: it does not disable the use of FACTER_url,
and it does not require changes to device modules, since it sets url internally.

The 'device()' method is valuable for use by tasks that need to target a device.