(PUP-8028) Ensure that directories are ignored by the file based loader
This commit changes the file based loader so that rejecs the directory
entries that are returned by the ruby `Dir#glob` method in case the
glob is executed without any known extension (as is the case when doing
a glob to find tasks).

The change fixes the problem where a directory would be interpreted as
a task, or interfere with a task using the same name (but different
extension) as the directory.