(PUP-8046) Fix problem with Ruby generator and Ruby keywords
When generating Ruby code from an `Object` data type containing
attributes or functions with names that are Ruby keywords, those names
were used verbatim in the generated code. This made it impossible to use
attribute or function names such as `ensure`.

This commit alters the generator so that all names that are reserved in
Ruby get prefixed with an underscore letter in the generated Ruby code.
The change is invisible from the Puppet code.