(PUP-7807) cleanup/add tests for aix remount options
Building on prior work, this commit DRYs up the code in mount a bit, sharing
behavior between bsd and aix, and breaking out a couple utility methods. A test
is added to assert that on AIX, if remounts is true and options are specified,
we pass the specified options with the appropriate 'remount' command.

Signed-off-by: Moses Mendoza <moses@puppet.com>