(PUP-8096) Preserve transactionstore when tags are applied
The transactionstore.yaml file used to calculate corrective change is updated on each agent run with the set of applied resources. When the --tags option is provided, only the state of the resources matching the tags were stored, with the state for other resources discarded. That means corrective change is not correctly determined on future runs for those resources excluded by the tags.

This commit updates the transactionstore when resources are skipped to preserve the last seen value.