(PUP-8008) Request module release listings from the Forge sorted by version
If there is a problem traversing additional pages of results from the
Forge, this at least will ensure that the latest versions of the
requested modules are found and installable. This also enables a future
optimization of only fetching the first page of results for each module
and seeing if that is sufficient to resolve dependencies.