(PUP-8078) Make it possible to create parameterized Object data types
This commit adds `type_parameters` as a top level entry to the `Object`
data type declaration. A type parameter is similar to an attribute but
only `type` and `value` is allowed in the declaration. As the name
suggests, this provides a way to declare how an `Object` data type
can be parameterized.  A new class `PObjectTypeExtension` is also added.
This class, together with the new `type_parameters` entry, forms the
mechanism that captures parameterized `Object` types.

A declaration like `T[P1,...Pn]` for an `Object` type `T` can be used
provided that `T` is declared with `type_parameters` that matches
`P1...Pn`. The declaration is captures using an instance of class
`PObjectTypeExtension` that appoints `T` and contains values for each
of the parameters.

The `PObjectTypeExtension` provides a default matching between the
values for the `type_parameters` of its base type `T` and instances of
`T`. The match will assume that any type parameter that is named the
same as an attribute should be used as a matcher for that attribute.
Matching is done, either by using instance matching of types (in case
the parameter is a type), by using regexps (in case the parameter is a
Regexp) or by using equality comparison.

A `PuppetObject` for which the type is a parameterized type, will infer
itself to a `PObjectTypeExtension` that contains parameters for which
a matches could be found using attribute names and values.

The default behavior for matching and inferring can be altered by
providing a special `PObjectTypeExtension` implementation written in
Ruby and added to the implementation registry. Such an impementation is
assumed to have the same name as the type that it extends with the
added suffix `TypeExtension`.