(PUP-8078) Fix glitch in how FromDataConverter resolves local references
Under some circumstances, the FromDataConverter would resolve a local
reference to a value in the same parent incorrectly since that reference
was not reachable in the current scope. This commit takes care of this
problem by changing how local references are resolved.

Local references is currently not in use by Puppet but are important
when more complex structures are serialized using tabulation.