(PUP-8078) Use case expression semantics when matching type parameters
This commit changes the way the values of an `Object`'s type parameters
are matched against the attribute values of an instance of the type or
against an other types type parameter values so that the match is
performed by the `CompareOperator` class. This means that the matching
is performed the same way as when evaluating a case expression.