(PUP-4647) yum provider uses rpm to uninstall packages
The yum provider uses its own install method (with a yum command) when
installing, but uses the inherited uninstall method from the rpm provider
(with a rpm command) when uninstalling.

This has the potential to 'corrupt' the yum database, reported as either:

rpmdb checksum is invalid
rpmdb altered outside of yum

With this commit, the yum provider also uses the yum command for uninstall.