(PUP-8107) Enable agent to specify source address
In an environment with multiple IP addresses on a system, there are
cases where the default system-selected source IP address may not be
appropriate for reaching the Puppet server due to access-lists, policy,
etc. This adds the `--sourceaddress <ADDRESS>` option to the CLI
allowing the user to supply the desired source IP address or hostname
for outbound connections.

Add check if Net::HTTP supports local_host (Ruby 2.x) and raise an
ArgumentError when unsupported.