(PUP-3323) Ensure we don't check gpg
When testing the yum utilities, we are creating an empty rpm package so
that we can install from a repo without hitting any external repos.
However, we aren't bothering to sign that empty package. When we go to
do the actual install, yum bails out because the package isn't signed.
This is great, but not in this instance. We don't care that the package
isn't signed, we just want to make sure the install works. This commit
adds in a flag to the install to make sure yum doesn't check the
signature.