(PUP-8078) Remove ability to declare default value for type parameters
This commit removes the `value` attribute from the `type_parameters`
declaration, effectively disabling the ability to use type parameter
defaults.

A type parameter that is given a value using a parameter is always used
as a constraint. When no value is given, the type parameter will not be
involved at all.

The `default` keyword can be used to represent an unset parameter. This
makes sense since the default behavior for any type parameter is to not
be involved. The keyword is needed when positional arguments are used
and parameters that are not last should remain unset. Using `undef` in
that situation doesn't work since it's a valid value in this context.