(PUP-8102) refactor the PAL API
This refactors the PAL API while retaining backwards compatibility. (No
changed tests in this commit).
The refactoring is performed to break up the initialization in two
steps:
* first create the environment
* then create a compiler

This is required in order to get performance from performing multiple
small steps inside one compiler while also making it possible to run a
sequence of independent compile operations.