(PUP-8078) Make parameterized type assignable when parameters are
This commit ensures that two parameterized types are assignable if their
base types, and all parameter values are assignable. This was not the
case earlier. If the value of a parameter type was a type, it just
matched values that were instances of that type or exactly equal to that
type.

Note: This only affects assignability between types. It does not affect
the check for if a value is an instance of the type.