(PUP-8114) Make enabled state check for SLES 11 boot services accurate in redhat service provider
Without this patch applied, the redhat service provider is not idempotent
in enabling SLES 11 boot services. These are special Sys V services beginning
with the string 'boot.'. When `chkconfig` is run against them, they print
'<name>  B' instead of '<name>  on', so the existing regex does not match them.
Thus the redhat provider always sees them as disables, and tries to enable
them on each agent run.

This patch adds the '<name>  B' condition to the regex.