(maint) Remove Configurer::DownloaderFactory
This commit removes the DownloaderFactory, which was an example of
unnecessary factory pattern paired with unused dependency injection.
This updates the Configurer constructor to no longer accept a factory as
an argument, which should be a private API change. If this turns out to
not be the case, this commit can be reverted.