(PUP-8102) Refactor PAL#with_script_compiler wrt code/manifest
Before this, a user would need to know that code_string: 'undef' must be
used to override the manifest and code settings. This is not something
users of the PAL API should need to know.

Now, a new option `configured_by_env` (defaults to false) gives the user
control over if the settings should be used, or if initial code is given
as args in the call. This also makes it clear that if nothing is given
then there is no initial code to load and evaluate.

Ergo, now it follows "law of least surprise".