(PUP-8121) Make #new function signatures more strict
The `Init` data type bases assignability and instance computation on the
signatures of the #new function for a given type. Before this commit,
those signatures were very relaxed. An `Integer` could for instance be
created from any `String` which gave that "blue" was an instance of
`Init[Integer]`. This commit tightens the constraints.