(PUP-8102) Refactor call_function and new_obj (+ rename)
This renames new_obj to create, and both create and calL_function now
takes a variable number of arguments instead of requiring them to be in
an array.

This is better because it is anticipated that most calls are going to be
done with arguments written inline instead of being in a variable. With
the changed style code is more readable.