(PUP-6739) Certain apps shouldn't fail if env is not local
Prior to this commit, we had a number of users running into an issue
where if they set the environment in either the toplevel or [main] of
puppet.conf, and that environment did not have a corresponding directory
in the local file system, commands like `puppet config` and `puppet
help` would bail out. This commit adds in the ability to specify
environment requirements on the application level. Commands like `puppet
config` and `puppet help` shouldn't require the environment to exist
locally.

This issue mainly surfaces because the concept of environment is used in
multiple contexts. The use case this change is meant to address is when
a user has set environment to refer to an environment that only exists
on the master. Under this scenario, `puppet agent` can check into the
master to request that specified environment. Although this change does
allow the user to continue to abuse how and where environment is set,
this is a common enough behavior we should have this buffer. This is
also useful if the user has accidentally mistyped their environment. If
this is the case, the user can still use `puppet config set` to update
that setting to the correct one.

For these cases, subcommands will default to the remote environment if
the specified environment does not exist locally.

This does not impact applications that should depend on the environment
existing locally, such as `puppet apply` and `puppet module`.