(PUP-8134) Allow tasks to specify that parameters are "sensitive"
This commit adds the task metadata field "sensitive" to the whitelist of
metadata fields. This field can be used to specify that a parameter
value should be redacted from logs, etc.

This does not include any behavior change to actually redact data, just
to allow those tasks to be run.