(PUP-8137) Make run_script accept arguments
Before this the run_script function did not accept arguments and did not
call bolt executor correctly.

Now, the run_script function accept a hash with a single key 'arguments'
that can be set to an array of strings to give as arguments to the
script. This solution was decided to make the run_script and run_task
functions similar in signature. It would otherwise be difficult to
support the list of targets in the same manner.