(PUP-8128) Add API allowing data types to dispatch method calls
When evaluating a call using dot notation, the evaluator would check
if the LHS of the dot expression (the receiver) was a PObjectType, and
if so, check if it had a member that corresponded to the RHS of the dot
expression. This approach did not allow other data types to dispatch
similar calls.

This commit adds a more general API for this mechanism where the
evaluator instead checks if the receiver is a `TypeWithMembers`. If
it is, it uses the access operator `[]` to retrieve the member which
then must implement `InvocableMember`. If a member is found, it will
be called.

The `PObjectType` and `PObjectTypeExtension` classes both implement
`TypeWithMembers`. The `PObjectType` attributes and functions implement
the `InvocableMember`.

The implementation will initially be used to gain access to attr_readers
of the `URI` type.