(PUP-8140) Add inspect method to Puppet::Face
Prior to this patch, error messages raised by `Puppet::Face` objects would
print the generic Ruby representation of the object, which is just the
class name plus the object instance id. The instance id is exactly as
useful as a random number when trying to figure out which face or version
of a face was involved in the error. This patch aliases the `inspect` method
of the `Puppet::Face` class to the `to_s` method so that error messages
include both the face name and version number.