(PUP-7401) Include type parameter in type mismatch report
Before this commit, a type mismtach involving the the types `Optional`,
`NotUndef`, `Sensitive`, and `Type` would be reported without the type
parameter when the base type differered between actual and expected
type. The resulting output was often somewhat cryptic since it's almost
always desirable to know the the contained type.

This commit ensures that the aformentioned types are always reported
with their included type.