(PUP-6739) Encourage users to limit where they set environment
This commits adds in a warning to encourage users to set the environment
in a subsection of puppet.conf rather than globally. This is because the
environment setting has different meaning depending on which context it
is being used in. It can either refer to the environment defined locally
on the filesystem or the environment that it is requesting from the
master. These different uses can lead to issues when the user has set
environment globally to an environment that only exists on the master
and not the agent. In this case, puppet subcommands that interpret
environment to refer to the environment as defined locally will fail
because that environment does not exist on the local filesystem. This
can be very frustrating and unintuitive to users.