(docs) Provide instructions on running puppet unit tests
This commits adds in a few details on exactly how to run all the tests.
Prior to this commit, we simply had a mandate that users run tests
before opening pull requests. This commit details how to do that. Users
are only expected to ensure the unit tests are passing prior to opening
a pull request.