(docs)(PUP-6739) Expand environment setting doc.
See [puppet-docs PR #816](https://github.com/puppetlabs/puppet-docs/pull/816)
by @melissa.

This commit expands the information provided for just how environment is
used when defined in puppet.conf. This variable has different
implications when defined in different contexts. Prior to this commit,
we simply had a very vague description of how it is used, with no
explicit detail about what happens if the user sets environment in the
different sections of puppet.conf. This commits adds in some of that
detail that was previously lacking.