(PUP-8170) Add support for `evaluate_literal` given an AST
This makes it possible to first parse and then use the resulting AST as
a literal value; for instance reading a string from the command line or
file and then interpreting it as a literal/static value of type string,
integer, float, boolean, regexp, array or hash.

An error is raised if the AST does not represent a literal value.