(PUP-8143) Pass "child" to ctrun on Solaris instead of "none"
In Puppet 5, some Solaris services were failing to start because Puppet
was abandoning them before making sure that they started correctly. This
caused Puppet to report that the service had started, when it actually
had not. Passing "child" to the ctrun command causes Puppet to manage
the service start more closely, while still retaining the contract
behavior for which ctrun was added (see PUP-2455).