(PUP-7507) Ensure that --facts are propagated to top scope and facts
Prior to this commit, the content of a yaml or json file passed to the
lookup CLI with the --facts option was just added to the node parameters
hash (which in essence means that they were just added to top scope).
This commit ensures that they are added to the contained facts instance
in the node and and hence, made available also when the dotted name
'facts.xxx' is used.