(PUP-8171) Add Pal::PlanSignature to be signature in the API
This adds a class Puppet::Pal::PlanSignature to act as a facade against
the implementation details related to the type/signature of a plan.

This is done so that a user of Pal does not have to know that a Plan is
a Function and that its signature is a Callable that you would need to
check if it can be called with arguments by name.

Now, this is instead done with a wrapper class "PlanSignature" that has
a method `callable_with?` that is easy to use.