(PUP-8171) Refactor the PAL API support for function signature
Before this, the PAL API had a method `function_signatures` that
returned an Array of PCallableType. This was somewhat difficult to use.

Now, the method is changed to `function_signature` and it returns an
instance of Puppet::Pal::FunctionSignature, a class that has the methods
`callable_with?`, and `callables`. The `callable_with?` method return
boolean true/false if the function can be called with a given set of
arguments, and it also accepts a block that will be called with a type
mismatch error message if args do not match. (This is consistent with
how the same is handled for a PlanSignature).

The method `signatures` returns an array of PCalableType if the user of
this API had the need for deeper digging into the signatures. This is
also consistent (in concept) with how this works in a PlanSignature
where a PStructType describing the parameters can be obtained.