(PUP-8013) Use different text domains for each environment
This commit adds the ability for an environment to set up a text domain
for the module translations associated with the environment. It will
only initialize the domain the first time it is used in the environment
lifecycle and just switch back to it each time it is needed thereafter,
reducing the amount of work that needs to be done to load translations
each time a catalog is requested for a different environment.