(PUP-8185) Make the PAL evaluating method instantiate definitions
Before this, when evaluating a string, file, or Program AST, the
evaluation method did not instantiate any definitions (functions, plans,
etc). This was bad because it is useful to be able to both define and
evaluate at the same time.

This is now fixed such that all evaluating methods instantiates the
definitiosn if asked to evaluate a Program. If evaluating arbitrary ast
(not a program) the definitions are not evaluated.