(PUP-8181) Rescue StopIteration for hash in each and map functions
Without this patch applied, the break() function fails on hash iterations
using each() or map(), with the error "break() from context where this is
illegal". break() is expected to break out of an iteration regardless of
context. This is happening because each() and map() rescue the StopIteration
exception for enumerables, but not for hashes. eval_Program() throws the
illegal context error when it catches the previously unhandled StopIteration
for a hash. This patch fixes the issue by rescuing StopIteration for hashes
in each() and map().

reduce() and slice() were already behaving as expected, while with() and
filter() are not conducive to a break() operation.