(PUP-8227) puppet apply/resource should not overwrite transactionstore
In the past enforcing change using "puppet apply" or "puppet resource"
would cause the transactionstore.yaml file to be overwritten. This
happened because persistence was enabled regardless of how
puppet was being run.

This commit adds the `enabled?` method to Transaction::Persistence which
now checks how puppet is being run to determine if persistence should be
enabled, which means we no longer overwrite transactionstore.yaml
during a "puppet apply" or "puppet resource" run.