(PUP-7894) Add Scheduled Task v2 API module
Previously Puppet used the V1 API for scheduled tasks however this has been
deprecated in favor of the V2 Win32OLE objects.  This commit adds a module which
can query and modify tasks with the V2 API.  This module only implements the
required methods for the existing Scheduled Task providers, for example it does
implement methods to run, terminate tasks or get last run information or exit
codes.