(PUP-5556) Make sprintf function support named arguments
This fixes a problem with using named arguments with sprintf (as claimed
to work in documentation). The issue is that the Ruby API requires
symbols to be used and there is no way to create symbols in the puppet
language (nor is it wanted).

The solution first tries a call to sprintf, and if it raises a KeyError
(because a named argument cannot be found due to the symbol/string
mismatch) then, a single hash argument is transformed to use symbol for
string keys, and a new attempt is made. (If that also fails, the
function will fail as the key is then really missing).