(maint) Suppress fuzzy translation warning
When an English string changes slightly, the FastGettext will
attempt to choose a translation that seems similar, update its English
key, and mark the translation as "fuzzy", to inform the translator that
review is needed. By default, FastGettext will issue a warning when a
fuzzy translation is used at runtime. This warning comes out of FastGettext
directly, circumventing our logging framework, and so is impossible for
a user to turn off, and in some cases impossible for them to fix.
Therefore, this commit passes a flag to FastGettext to suppress the
warning.