(PUP-8090) Load module translations when using a cached catalog
Before this commit, when using a cached catalog, module translations
did not get loaded because we do not pluginsync when using a cached
catalog. This commit ensures that translations get loaded when using
a cached catalog. The translations must be loaded before preparing the
catalog to ensure that messages from types get translated.