(PUP-8191) Don't initialize gettext for modules without relevant translations
This commit ensures that a module's locales directory contains a set of
translation files for the selected locale before creating the
translation repository for that module. Since translation files are
lazily loaded by FastGettext, it will create the repo first and look for
the files later. If it can't find them, it will create and read an
empty MO file as a placeholder. To avoid this extra work, we now do that
check up front.