(PUP-7894) Add basic integration tests for scheduled tasks
Previously the Scheduled Task API helpers did not have any tests to confirm
their behaviour.  This commit adds some basic tests to the Scheduled Task V2
module to ensure it can list, create, delete, and modify tasks on a Windows
system.  This commit also adds integration tests for the V2 facade class to
confirm that tasks created in the V1 API appear in the V2 facade and vice versa.