(PUP-8171) Address comments from review
This changes a typo in a comment and renames a variable `block` that
could easily be confused with a `&block` since the method also yields to
a given block. The parameter is renamed to `callable` (other useful
words are already keywords or confusing in a different way; lambda,
proc, etc).