(PUP-8202) Add TaskSignature, and refactor FunctionSignature
This adds a TaskSignature class that can answer if a task can be run
with a given hash of arguments. This also refactors FunctionSignature to
take a Function Class instead of a Function instance. That was done to
allow it to use the new-function Class from the new-function for a task
in its implementation of checking hash arguments.