(PUP-8204) Refactor so that valid_path is determined by SmartPath
This commit refactors the way paths are found when using a `SmartPath`
to search the cache that is maintained by a `FileBased` loader so that
the `SmartPath` is responsible for the decision whether or not a path
is valid.

Moving this decision into the `SmartPath` makes it possible for the
task specific smart-path to only allow paths that represent valid tasks
and also reject forbidden extensions.