(PUP-8204) Add EmptyLoader class and use it for static environment
A static environment doesn't have a path and using FileBased loader
resulted in problems because the lack of a path made the loader find
things under the 'tasks' folder in the puppet lib directory. This commit
adds a new `EmptyLoader` which is used in case the environment has no
path.