(PUP-8090) Do not reference the Locale gem outside of GettextConfig
The GettextConfig module controls all access to the FastGettext and
Locale libraries. A previous commit accidentally introduced a reference
to the Locale library outside of this module, which caused errors during
puppetserver unit testing, when the gem was not installed. This commit
moves all references to Locale behind the guards set up in the
GettextConfig module.