(PUP-8212) Fix strange behavior of empty Variant data type
This changes the (broken) type algebra for empty Variant type such that
it is only assignable to empty Variant, Any, NotUndef[Variant], and
Unit, and that any Variant is assignable to an empty Variant.

This modifies the setup in `types_setup` by removing Variant type from
`all_types` - this because it is not meaninful to test all other data
types against empty Variant with one and the same rule. Instead, there
are now explicit tests for empty Variant.