(PUP-8191) Create PO file for "en" locale for Puppet
This commit adds an empty English PO file for Puppet. If FastGettext
can't find a file for the requested locale, it will repeatedly
open an empty "translation file" placeholder that it creates
on the fly, every time it tries to look up a translation. With an actual
file in place, the file will be loaded once when the tooling is set up,
and the rest of the machinery will work as intended, checking the cached
result of that load for translations whenever one is requested.  When
FastGettext cannot find a requested string among its loaded translations,
it will fall back to what's in source. The file for English added in
this commit is therefore empty, since it behaves the same as one will
full English "translations", but we don't need to worry about keeping it
up to date when strings change in Puppet.

We considered instead trying to stub or disable translation loading
entirely for English in particular, but rejected this route because we
didn't want to create more speical cases that could hide performance
issues for international users.