(PUP-8199) Use instances of Task instead of different Task sub-types
Prior to this commit, each task that was described with metadata became
a type of its own in the puppet type system. An instance of such a task
would represent a task invocation, complete with settings and parameter
values. This was problematic, primarily because the parameters were
represented by attributes and hence subject to the same naming rules as
the attributes, but also because it was cumbersome to filter out tasks
from other types in the type system.

This commit changes the way the `Task` type is used so that a task
description, instead of becoming a type in its own right, becomes an
instance of the one and only `Task` type. Creating an object that
represents a task invocation (description + parameter values) is no
longer possible.

The loader namespace `:task` is added to the Loaders. This namespace
must be used when searching for tasks. Using `:type` will now only find
the `Task` type.