(PUP-8124) Resolve qualified names in class parameter scope
A regression introduced in PUP-5635 caused the evaluator to omit the
parameter scope when evaluating parameter default expressions and
resolving qualified names of variables. As a result, a reference to a
parameter to the left of the one declaring the default expression,
declared using the qualified name of the class holding the parameter
list, would be unresolved.

This commit ensures that resolution of qualified names will include the
parameter scope when the name appoints the current class.