(maint) Drop dead 'bindings.rb'
Before this, the bindings.rb file that made it possible to register
bindings for the in puppet 5.x removed "bindings system"/injector
remained. This code was no longer used by anything in puppet, and while
it was possible to call it to register things, there was nothing at the
other end picking up such registrations - so, would not have had the
desire effect for anyone anyway.

This drops the 'bindings.rb' file and one require of this file.