(PUP-6043) Change syntax error "at end of file" to "end of input"
This changes the error message produced when there is a syntax error
because the parser hits the end of its input. When this happens the
parser does not know if the text it is parsing is from a file or from a
string, or from the command line.

Since users found it confusing that the error speaks of "end of file"
when there is no file involved, the best is thought to be naming it "end
of input" as that works for all of the various ways input may be given.