(PUP-5206) Verify content property uses clientbucketdir
The test didn't fail when applying a file resource whose content referenced a
filebucket checksum, because the file was already in-sync. Modifying the file
forces the apply run to remediate the file, and for that to succeed it needs to
be using the same local bucket (aka `clientbucketdir`) that was used to backup
the file in the first place.