(PUP-8229) Assume that Puppet 'lib' directory can be relocated
A regression in PUP-7864 add the assumption that the 'lib' directory in
puppet would always be present. This assumption is false since the 'lib'
directory is renamed to 'site_ruby' during installation.

There are no unit tests to check this (since a bundle exec always has
a 'lib' under the puppet root). The problem, if present, is however
caught by acceptance tests that uses the loader.