(PUP-6162) Improve error message when local type is faulty in func
When a local type is faulty inside a 4.x ruby function the error message
did not in any way reveal where the problem was. It may surface
something like "syntax error when calling the function 'foo'".

Now it mentions that it is an error while loading the function, the name
of the function, and if the error is a parse error from parsing a data
type specification, it includes the source and location within that
string.