(PUP-8232) Update facts upload face
This commit updates the facts upload face to use settings from the agent
section of puppet.conf and to retrieve facts using node_name_value and
node_name_fact instead of certname to match behavior with `puppet agent` and
`puppet apply`. The name of the server to which facts were sent is also logged.

Finally, documentation of the face has been updated with a link to current
auth.conf docs.