(PUP-8237) Simplify function parameter injection
This commit removes some obsolete logic that was designed to handle
injection performed by the advanced but not really used injection system
that was earlier removed from Puppet core.

In essence, instead of passing a four element array with various data
about the injection, only a symbol denoting the parameter to be injected
is passed on. Dead code in the form of the two methods `injected_param`
and `injected_producer_param` is removed.