(PUP-8237) Make Puppet::Pal::ScriptCompiler available to functions
This commit makes the `Puppet::Pal::ScriptCompiler` available using a
`Puppet::Context` override. The compiler is available using the symbol
`:pal_script_compiler`.

A new special parameter type named `script_compiler_param` is also added
to the Ruby function dispatcher builder for `InternalFunction`. Usage is
similar to `scope_param` but it injects the script compiler instead of
injecting the scope.