(PUP-8244) Call the accessor instead of assigning to local variable
Previously we were trying to call accessor methods, but ruby will
instead create a local variable that shadows the setter[1]:

    When using method assignment you must always have a receiver. If you do
    not have a receiver Ruby assumes you are assigning to a local variable.

Puppet::Resource removes the call because the environment attribute is
invoked in the attributes.each block below. Note in the case that an
environment attribute doesn't exist, we never call
`send('environment=')`. However that is consistent behavior with what
happened previously, since `environment = nil` was a noop.

Puppet::Resource::Status removes the line, because the `failed` instance
variable is already set when adding the event with a `failure` status.

[1] https://docs.ruby-lang.org/en/2.0.0/syntax/assignment_rdoc.html