(PUP-7394) Change behavior of replace_file to change file mode after yield
Previously, when replacing an existing file, replace_file would set the temp
file's permissions to be that of the existing file or the provided default
permissions before yielding to the block. This was done originally because the
temp file was created with insecure file permissions. This is no longer
necessary, because Puppet::FileSystem::Uniquefile creates temporary files with
mode 0600.

The previous behavior of using the existing file's mode before yielding to the
block meant that we couldn't write to the temporary file if the file we were
trying to replace was read-only. By setting the final permissions of the file
after yielding to the block we no longer run into this issue.